% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_match.R
\name{direct_match}
\alias{direct_match}
\title{Direct Match}
\usage{
direct_match(df, target_df = NULL)
}
\arguments{
\item{df}{A \code{tibble} containing the species data to be matched.}

\item{target_df}{A \code{tibble} representing the ppendemic database containing the reference list of endemic species.}
}
\value{
A \code{tibble} with an additional logical column \code{direct_match}, indicating whether the binomial or trinomial name was successfully matched (\code{TRUE}) or not (\code{FALSE}).
}
\description{
This function performs a direct match of species names. It matches the genus and species if the name is binomial, and matches the genus, species, and infraspecies if the name includes a subspecies.
}
\keyword{already}
\keyword{and}
\keyword{backbone}
\keyword{because}
\keyword{columns}
\keyword{compatilbility}
\keyword{empty}
\keyword{ensure}
\keyword{exists}
\keyword{for}
\keyword{input}
\keyword{internal}
\keyword{issue}
\keyword{needed}
\keyword{of}
\keyword{second}
\keyword{sequential_matching:}
\keyword{solve}
\keyword{the}
\keyword{there}
\keyword{tibble,}
\keyword{to}
\keyword{with}
