% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_match_spp_genus.R
\name{direct_match_species_within_genus_helper}
\alias{direct_match_species_within_genus_helper}
\title{Direct Match Species within Genus}
\usage{
direct_match_species_within_genus_helper(df, target_df)
}
\arguments{
\item{df}{A tibble containing the species data to be matched.}

\item{target_df}{A tibble representing the ppendemic database containing the reference list of endemic species.}
}
\value{
A tibble with an additional logical column indicating whether the specific epithet was successfully matched within the matched genus (\code{TRUE}) or not (\code{FALSE}).
}
\description{
This function performs a direct match of specific epithets within an already matched genus from the list of endemic species in the ppendemic database.
}
\keyword{internal}
