% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{especiespp}
\alias{especiespp}
\title{endemic plants list}
\format{
An object of class \code{character} of length 5408.
}
\usage{
especiespp
}
\description{
endemic plants list
}
\keyword{datasets}
