% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_match_infraspp_within_species.R
\name{fuzzy_match_infraspecies_within_species}
\alias{fuzzy_match_infraspecies_within_species}
\title{Fuzzy Match Infraspecies within Species}
\usage{
fuzzy_match_infraspecies_within_species(df, target_df = NULL)
}
\arguments{
\item{df}{A tibble containing the species data to be matched.}

\item{target_df}{A tibble representing the ppendemic database containing the reference list of endemic species.}
}
\value{
A tibble with an additional logical column fuzzy_match_infraspecies_within_species, indicating whether the specific infraspecies was successfully fuzzy matched within the matched species (\code{TRUE}) or not (\code{FALSE}).
}
\description{
This function performs a fuzzy match of specific infraspecies within an already matched epithet from the list of endemic species in the ppendemic database.
}
\keyword{internal}
