% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genus_match.R
\name{genus_match}
\alias{genus_match}
\title{Match Genus Name}
\usage{
genus_match(df, target_df = NULL)
}
\arguments{
\item{df}{A \code{tibble} containing the genus names to be matched.}

\item{target_df}{A \code{tibble} representing the ppendemic database containing the reference list of endemic species.}
}
\value{
A \code{tibble} with an additional logical column \code{genus_match}, indicating whether the genus was successfully matched (\code{TRUE}) or not (\code{FALSE}).
}
\description{
This function performs a direct match of genus names against the genus names listed in the ppendemic database.
}
\keyword{internal}
