% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-to-file.R
\name{path_to_file}
\alias{path_to_file}
\title{Get file path to \code{ppendemic} files}
\source{
This function is adapted from \code{readxl::readxl_example()}.
}
\usage{
path_to_file(path = NULL)
}
\arguments{
\item{path}{\code{NULL}, the example files will be listed.}
}
\description{
ppendemic comes bundled with csv files in its \code{inst/extdata}
directory. This function make them easy to access.
}
\examples{
path_to_file()

}
