% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep_by_reg.R
\name{pep_by_reg}
\alias{pep_by_reg}
\title{pep_by_reg}
\usage{
pep_by_reg(...)
}
\arguments{
\item{...}{Atomic element or vector with species names.}
}
\value{
A tibble with two variables: species name and regions where species found
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

List all endemic species within an region.
}
\examples{
# Basic usage
### One name
pep_by_reg("Aphelandra latibracteata")
### Set of names
species <- c("Sanchezia filamentosa", "Aphelandra latibracteata")
pep_by_reg(species)
}
