% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep_regdep.R
\name{pep_regdep}
\alias{pep_regdep}
\title{pep_regdep}
\usage{
pep_regdep(...)
}
\arguments{
\item{...}{Atomic element or vector with species names.}
}
\value{
A tibble with two variables: species name and regions where species found
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

Help to find in which region the endemic species are described
}
\examples{
# Basic usage
### One name
pep_regdep("Aphelandra latibracteata")
### Set of names
species <- c("Sanchezia filamentosa", "Aphelandra latibracteata")
pep_regdep(species)
pep_regdep("Sanchezia filamentosa", "Aphelandra latibracteata")
}
