% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep_regecol.R
\name{pep_regecol}
\alias{pep_regecol}
\title{pep_regecol}
\usage{
pep_regecol(region)
}
\arguments{
\item{region}{Atomic element or vector with ecological region names,
shrot or long name.}
}
\value{
A tibble
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

List all endemic species within an ecological region.
}
