% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_poorman.R
\name{pipe}
\alias{pipe}
\alias{\%>\%}
\title{Forward-pipe operator}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs}{The result you are piping.}

\item{rhs}{Where you are piping the result to.}
}
\description{
Pipe an object forward into a function or call expression,
from \href{https://nathaneastwood.github.io/poorman/}{poorman}
}
\author{
Nathan Eastwood and Antoine Fabri \email{antoine.fabri@gmail.com}.
}
