% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppendemic-package.R
\docType{package}
\name{ppendemic-package}
\alias{ppendemic}
\alias{ppendemic-package}
\title{ppendemic: The red book of endemic plants of Peru}
\description{
Provides tools for accessing the endemic plant's information from  "The red book of endemic plants of Peru ".  The ppendemic package contains a database with cleaned and standardized botanical data including occurrence and taxonomic data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PaulESantos/ppendemic/}
  \item Report bugs at \url{https://github.com/PaulESantos/ppendemic/issues}
}

}
\author{
\strong{Maintainer}: Paul Efren Santos Andrade \email{paulefrens@gmail.com}

Authors:
\itemize{
  \item Lucely L. Vilca Bustamante \email{vilca.lu01@gmail.com}
}

}
\keyword{internal}
