% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regiones_ecologicas.R
\docType{data}
\name{regiones_ecologicas}
\alias{regiones_ecologicas}
\title{Specie name and ecological region for plants listed in El  libro
rojo  de  las  plantas  endémicas  del  Perú}
\format{
A tibble with 6778 rows and 5 variables:
\describe{
\item{accepted_family}{a factor denoting plant family}
\item{accepted_genus}{a factor denoting plant genus}
\item{accepted_name}{a factor denoting plant species}
\item{region_id}{a factor denoting short name for ecological
region}
\item{region_eco}{a factor denoting ecological region names.
Based on: "ZAMORA, C. 1996. Las regiones ecológicas del Perú.
En: Rodríguez, L.O. (ed.), Diversidad Biológica del Perú: Zonas
Prioritarias para su Conservación, pp. 137-141. FANPE,
GTZ-INRENA. Lima, Perú." }
}
}
\source{
{Originally published in: El  libro  rojo  de  las  plantas
endémicas  del  Perú (2006).  Ed.:  Blanca  León  et al.Rev.  peru.
biol.  Número  especial  13(2)}
}
\usage{
regiones_ecologicas
}
\description{
Specie name and ecological region for plants listed in El  libro
rojo  de  las  plantas  endémicas  del  Perú
}
\keyword{internal}
