% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registro_departamental.R
\docType{data}
\name{registro_departamental}
\alias{registro_departamental}
\title{Species names and region where finded from El  libro  rojo  de  las  plantas  endémicas  del  Perú}
\format{
A tibble with 9041rows and 2 variables:
\describe{
\item{accepted_name}{a factor denoting plant species}
\item{registro_dep}{a factor denoting region names}
}
}
\source{
{Originally published in: El  libro  rojo  de  las  plantas  endémicas  del  Perú (2006).  Ed.:  Blanca  León  et al.Rev.  peru.  biol.  Número  especial  13(2)}
}
\usage{
registro_departamental
}
\description{
Species names and region where finded from El  libro  rojo  de  las  plantas  endémicas  del  Perú
}
\keyword{datasets}
