% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suffix_match_spp_within_genus.R
\name{suffix_match_species_within_genus_helper}
\alias{suffix_match_species_within_genus_helper}
\title{Suffix Match Species within Genus}
\usage{
suffix_match_species_within_genus_helper(df, target_df)
}
\arguments{
\item{df}{a \code{tibble}.}

\item{target_df}{A \code{tibble} representing the ppendemic database containing the reference list of endemic species.}
}
\value{
Returns a \code{tibble} with the additional logical column
\code{suffix_match_species_within_genus}, indicating whether the specific
epithet was successfully matched within the matched genus (TRUE)
or not (FALSE).
}
\description{
Function to match the specific epithet by exchanging common suffixes
within an already matched genus into the ppendemic database.
The following suffixes are captured: \code{c("a", "i", "is", "um", "us", "ae")}
}
\keyword{"ae",}
\keyword{"escens")}
\keyword{"i",}
\keyword{"is",}
\keyword{"oides",}
\keyword{"um",}
\keyword{"us",}
\keyword{<-}
\keyword{and}
\keyword{both}
\keyword{c("a",}
\keyword{column}
\keyword{common_suffixes}
\keyword{create}
\keyword{database}
\keyword{in}
\keyword{input}
\keyword{internal}
\keyword{root}
\keyword{subset}
\keyword{the}
\keyword{user}
\keyword{word}
