% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_match.R
\name{direct_match}
\alias{direct_match}
\title{Direct Match}
\usage{
direct_match(df, target_df = NULL)
}
\arguments{
\item{df}{A tibble containing the species data to be matched.}

\item{target_df}{A tibble representing the ppendemic database containing the reference list of endemic species.}
}
\value{
A tibble with an additional logical column direct_match indicating whether the binomial or trinomial name was successfully matched (\code{TRUE}) or not (\code{FALSE}).
}
\description{
This function performs a direct match of species names. It matches the genus and species if the name is binomial, and matches the genus, species, and infra species if the name includes a subspecies.
}
\keyword{internal}
