% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_match_genus.R
\name{fuzzy_match_genus}
\alias{fuzzy_match_genus}
\title{Fuzzy Match Genus Name}
\usage{
fuzzy_match_genus(df, target_df = NULL)
}
\arguments{
\item{df}{A \code{tibble} containing the genus names to be matched.}

\item{target_df}{A \code{tibble} representing the ppendemic database containing the reference list of endemic species.}
}
\value{
A \code{tibble} with two additional columns:
\itemize{
\item \code{fuzzy_match_genus}: A logical column indicating whether the genus was successfully matched (\code{TRUE}) or not (\code{FALSE}).
\item \code{fuzzy_genus_dist}: A numeric column representing the distance for each match.
}
}
\description{
This function performs a fuzzy match of genus names against the ppendemic database using \code{fuzzyjoin::stringdist()} to account for slight variations in spelling.
}
\keyword{(present}
\keyword{Alternatively}
\keyword{In}
\keyword{cases}
\keyword{choose}
\keyword{choosing}
\keyword{could}
\keyword{databases)}
\keyword{first}
\keyword{for}
\keyword{here:}
\keyword{in}
\keyword{instance}
\keyword{internal}
\keyword{like}
\keyword{match.}
\keyword{matches:}
\keyword{more}
\keyword{multiple}
\keyword{of}
\keyword{one}
\keyword{something}
\keyword{sophisticated}
\keyword{support}
\keyword{the}
\keyword{use}
\keyword{we}
\keyword{with}
