% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_match_genus.R
\name{fuzzy_match_genus}
\alias{fuzzy_match_genus}
\title{Fuzzy Match Genus Name}
\usage{
fuzzy_match_genus(df, target_df = NULL)
}
\arguments{
\item{df}{A tibble containing the genus names to be matched.}

\item{target_df}{A tibble representing the ppendemic database containing the reference list of endemic species.}
}
\value{
A tibble with two additional columns:
\itemize{
\item fuzzy_match_genus: A logical column indicating whether the genus was successfully matched (\code{TRUE}) or not (\code{FALSE}).
\item fuzzy_genus_dist: A numeric column representing the distance for each match.
}
}
\description{
This function performs a fuzzy match of genus names against the ppendemic database using fuzzyjoin::stringdist() to account for slight variations in spelling.
}
\keyword{internal}
