% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep_check.R
\name{pep_check}
\alias{pep_check}
\title{pep_check}
\usage{
pep_check(x)
}
\arguments{
\item{x}{Vector}
}
\value{
vector
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

This function allows evaluating whether or not the species is endemic to Peru.
}
\examples{
# Basic usage
spp <-  c("Hedyosmum", "Miconia alpina", "Persea sp2")
pep_check(spp)


}
