% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep_check.R
\name{pep_check}
\alias{pep_check}
\title{pep_check}
\usage{
pep_check(x)
}
\arguments{
\item{x}{Vector conteniendo el listado de especies.}
}
\value{
vector
}
\description{
Esta función permite evaluar si la especie es o no endémica del Perú.
}
\examples{
# Basic usage
spp <-  c("Hedyosmum", "Miconia alpina", "Persea sp2")
pep_check(spp)


}
