% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep_check.R
\name{pep_check}
\alias{pep_check}
\title{pep_check}
\usage{
pep_check(x)
}
\arguments{
\item{x}{Vector conteniendo el listado de especies.}
}
\value{
vector
}
\description{
Esta función permite evaluar si la especie es o no endémica del Perú.
}
\examples{

spp <-  c("Clethra cuneata", "Miconia setulosa", "Hedyosmum"
          "Weinmannia fagaroides", "Symplocos quitensis" , "Miconia alpina",
          "Persea sp2", "Myrsine", "Symplocos var",
          "Polylepis pauta var pauta", "Senna versicolor var. heterosperma")
pep_check(spp)


}
