% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep_check.R
\name{pep_check}
\alias{pep_check}
\title{pep_check}
\usage{
pep_check(x)
}
\arguments{
\item{x}{an atomic character string or a vector with the species names that will be verified if it endemic species}
}
\value{
a vector with information for each species names tested: a) "endemic" if the species is endemic of Perú; b)  "not endemic" if the species is not endemic of Perú; c)"not binary name" if only the genera were submitted; d) "taxon state undefined" if the submitted species name has a taxonomic status undetermined.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
The function automates the process to check if a plant species is or not endemic to Perú.
}
\examples{
# Basic usage
spp <-  c("Hedyosmum", "Miconia alpina", "Persea sp2")
pep_check(spp)


}
