% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep_regdep_map.R
\name{pep_regdep_map}
\alias{pep_regdep_map}
\title{pep_regdep_map}
\usage{
pep_regdep_map(spp_name)
}
\arguments{
\item{spp_name}{species name}
}
\value{
ggplot object
}
\description{
pep_regdep_map
}
\examples{

pep_regdep_map("Grosvenoria coelocaulis")
}
