% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep_regdep_map.R
\name{pep_regdep_map}
\alias{pep_regdep_map}
\title{pep_regdep_map}
\usage{
pep_regdep_map(spp_name)
}
\arguments{
\item{spp_name}{species name}
}
\value{
a map showing all regions where the species is present. This map is a ggplot object that could be edited.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function builds a map indicating where the species has been recorded.
}
\examples{
# Basic usage
pep_regdep_map("Grosvenoria coelocaulis")
}
