% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peru_shp.R
\docType{data}
\name{peru}
\alias{peru}
\title{For Peru at scale 1:10m (high).}
\format{
A \code{SpatialPolygonsDataFrame}

An object of class \code{sf} (inherits from \code{data.frame}) with 26 rows and 85 columns.
}
\usage{
peru
}
\description{
For Peru at scale 1:10m (high).
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame with attributes}
}}

\keyword{datasets}
