% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppendemic-package.R
\docType{package}
\name{ppendemic-package}
\alias{ppendemic}
\alias{ppendemic-package}
\title{ppendemic: The Red Book of Endemic Plants of Peru}
\description{
Gives help to access the information from  "The Red Book of Endemic Plants of Peru".  The package contains a database with cleaned and standardized botanical data including occurrence and taxonomic data. The original source could be found in "León, Blanca, Nigel Pitman, y José Roque. (2006). Introducción a Las Plantas Endémicas Del Perú"<doi:10.15381/rpb.v13i2.1782><https://revistasinvestigacion.unmsm.edu.pe/index.php/rpb/issue/view/153>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PaulESantos/ppendemic/}
  \item Report bugs at \url{https://github.com/PaulESantos/ppendemic/issues}
}

}
\author{
\strong{Maintainer}: Paul Efren Santos Andrade \email{paulefrens@gmail.com} (\href{https://orcid.org/0000-0002-6635-0375}{ORCID})

Authors:
\itemize{
  \item Lucely L. Vilca Bustamante \email{vilca.lu01@gmail.com}
}

}
\keyword{internal}
