% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regiones_ecologicas.R
\docType{data}
\name{regiones_ecologicas}
\alias{regiones_ecologicas}
\title{Specie name and ecological region for plants listed in El  libro  rojo  de  las  plantas  endémicas  del  Perú}
\format{
A tibble with 9041rows and 2 variables:
\describe{
\item{accepted_name}{a factor denoting plant species}
\item{region_id}{a factor denoting short name for ecological region}
\item{region_eco}{a factor denoting ecological region names}
}
}
\source{
{Originally published in: El  libro  rojo  de  las  plantas  endémicas  del  Perú (2006).  Ed.:  Blanca  León  et al.Rev.  peru.  biol.  Número  especial  13(2)}
}
\usage{
regiones_ecologicas
}
\description{
Specie name and ecological region for plants listed in El  libro  rojo  de  las  plantas  endémicas  del  Perú
}
\keyword{datasets}
